refactor(@remix-run/react): move serialize type utilities into their own module
