+++ packages/remix-react/components.tsx
@@ -20,7 +20,6 @@ import {
   useResolvedPath,
 } from "react-router-dom";
 import type { LinkProps, NavLinkProps } from "react-router-dom";
-import type { Merge } from "type-fest";
 import { createPath } from "history";
 
 import type { AppData, FormEncType, FormMethod } from "./data";
@@ -56,6 +55,7 @@ import type {
   Fetcher,
   Submission,
 } from "./transition";
+import type { UseDataFunctionReturn } from "./serialize";
 
 ////////////////////////////////////////////////////////////////////////////////
 // RemixEntry
@@ -1355,74 +1355,6 @@ export function useMatches(): RouteMatch[] {
     [matches, routeData, routeModules]
   );
 }
-
-export type TypedResponse<T> = Response & {
-  json(): Promise<T>;
-};
-
-type DataFunction = (...args: any[]) => unknown; // matches any function
-type DataOrFunction = AppData | DataFunction;
-type JsonPrimitive =
-  | string
-  | number
-  | boolean
-  | String
-  | Number
-  | Boolean
-  | null;
-type NonJsonPrimitive = undefined | Function | symbol;
-
-type IsAny<T> = 0 extends 1 & T ? true : false;
-
-// prettier-ignore
-type Serialize<T> =
-  IsAny<T> extends true ? any :
-  T extends JsonPrimitive ? T :
-  T extends NonJsonPrimitive ? never :
-  T extends { toJSON(): infer U } ? U :
-  T extends [] ? [] :
-  T extends [unknown, ...unknown[]] ? SerializeTuple<T> :
-  T extends ReadonlyArray<infer U> ? (U extends NonJsonPrimitive ? null : Serialize<U>)[] :
-  T extends object ? SerializeObject<UndefinedToOptional<T>> :
-  never;
-
-/** JSON serialize [tuples](https://www.typescriptlang.org/docs/handbook/2/objects.html#tuple-types) */
-type SerializeTuple<T extends [unknown, ...unknown[]]> = {
-  [k in keyof T]: T[k] extends NonJsonPrimitive ? null : Serialize<T[k]>;
-};
-
-/** JSON serialize objects (not including arrays) and classes */
-type SerializeObject<T extends object> = {
-  [k in keyof T as T[k] extends NonJsonPrimitive ? never : k]: Serialize<T[k]>;
-};
-
-/*
- * For an object T, if it has any properties that are a union with `undefined`,
- * make those into optional properties instead.
- *
- * Example: { a: string | undefined} --> { a?: string}
- */
-type UndefinedToOptional<T extends object> = Merge<
-  {
-    // Property is not a union with `undefined`, keep as-is
-    [k in keyof T as undefined extends T[k] ? never : k]: T[k];
-  },
-  {
-    // Property _is_ a union with `defined`. Set as optional (via `?`) and remove `undefined` from the union
-    [k in keyof T as undefined extends T[k] ? k : never]?: Exclude<
-      T[k],
-      undefined
-    >;
-  }
->;
-
-export type UseDataFunctionReturn<T extends DataOrFunction> = Serialize<
-  T extends (...args: any[]) => infer Output
-    ? Awaited<Output> extends TypedResponse<infer U>
-      ? U
-      : Awaited<Output>
-    : Awaited<T>
->;
 
 /**
  * Returns the JSON parsed data from the current route's `loader`.
